[implementing] [refactor "Preset" to "Goal", include database table and model]

+++ Programming/Nevo/app/src/main/java/com/medcorp/nevo/database/DatabaseHelper.java
@@ -13,7 +13,7 @@ import com.medcorp.nevo.R;
 import com.medcorp.nevo.database.dao.AlarmDAO;
 import com.medcorp.nevo.database.dao.HeartbeatDAO;
 import com.medcorp.nevo.database.dao.IDailyHistory;
-import com.medcorp.nevo.database.dao.PresetDAO;
+import com.medcorp.nevo.database.dao.GoalDAO;
 import com.medcorp.nevo.database.dao.SleepDAO;
 import com.medcorp.nevo.database.dao.StepsDAO;
 import com.medcorp.nevo.database.dao.UserDAO;
@@ -31,7 +31,7 @@ import java.sql.SQLException;
 public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
 
     private static final String DATABASE_NAME = "nevowatch.db";
-    private static final int DATABASE_VERSION = 3;
+    private static final int DATABASE_VERSION = 4; // from 3 to 4, refactor table "Preset" to "Goal"
 
     private  Dao<IDailyHistory,Integer> dailyhistoryDao = null;
     private  Dao<UserDAO,Integer> userDao = null;
@@ -39,8 +39,8 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
     private  Dao<StepsDAO,Integer> stepsDao = null;
     private  Dao<HeartbeatDAO,Integer> heartbeatDAO = null;
     private  Dao<AlarmDAO,Integer> alarmDao = null;
-    private  Dao<PresetDAO,Integer> presetsDAO = null;
-
+    private  Dao<GoalDAO,Integer> goalsDAO = null;
+    private  Context context;
 
     //Classic singleton
     private static DatabaseHelper instance = null;
@@ -56,6 +56,7 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
 
     private DatabaseHelper(Context context) {
         super(context, DATABASE_NAME, null, DATABASE_VERSION, R.raw.ormlite_config);
+        this.context = context;
     }
 
     @Override
@@ -67,7 +68,7 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
             TableUtils.createTable(connectionSource, StepsDAO.class);
             TableUtils.createTable(connectionSource, HeartbeatDAO.class);
             TableUtils.createTable(connectionSource, AlarmDAO.class);
-            TableUtils.createTable(connectionSource, PresetDAO.class);
+            TableUtils.createTable(connectionSource, GoalDAO.class);
         } catch (SQLException e) {
             Log.e(DatabaseHelper.class.getName(), "Unable to create datbases", e);
         }
@@ -76,13 +77,25 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
     @Override
     public void onUpgrade(SQLiteDatabase sqliteDatabase, ConnectionSource connectionSource, int oldVer, int newVer) {
         try {
+            //when user update app from v2.0.1(database is v3) to new version (database >= v4)
+            //here create new table "Goal" and add some default records, the best way is copy data from  table "Preset" to "Goal",but now I can't read the table "Preset"
+            //due to the PresetDAO class has been renamed to GoalDAO class.
+            if(oldVer == 3 && newVer >=4)
+            {
+                TableUtils.createTable(connectionSource, GoalDAO.class);
+                getGoalDao().createIfNotExists(new GoalDAO(context.getString(R.string.startup_goal_light), true, 7000));
+                getGoalDao().createIfNotExists(new GoalDAO(context.getString(R.string.startup_goal_moderate), true, 10000));
+                getGoalDao().createIfNotExists(new GoalDAO(context.getString(R.string.startup_goal_heavy), true, 20000));
+                //here return for keep user history data.
+                return;
+            }
             TableUtils.dropTable(connectionSource, IDailyHistory.class, true);
             TableUtils.dropTable(connectionSource, UserDAO.class, true);
             TableUtils.dropTable(connectionSource, SleepDAO.class, true);
             TableUtils.dropTable(connectionSource, StepsDAO.class, true);
             TableUtils.dropTable(connectionSource, HeartbeatDAO.class, true);
             TableUtils.dropTable(connectionSource, AlarmDAO.class, true);
-            TableUtils.dropTable(connectionSource, PresetDAO.class, true);
+            TableUtils.dropTable(connectionSource, GoalDAO.class, true);
             onCreate(sqliteDatabase, connectionSource);
         } catch (SQLException e) {
             Log.e(DatabaseHelper.class.getName(), "Unable to upgrade database from version " + oldVer + " to new "
@@ -123,10 +136,10 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
         return alarmDao;
     }
 
-    public Dao<PresetDAO, Integer> getPresetDao() throws SQLException {
-        if (presetsDAO== null)
-            presetsDAO= getDao(PresetDAO.class);
-        return presetsDAO;
+    public Dao<GoalDAO, Integer> getGoalDao() throws SQLException {
+        if (goalsDAO == null)
+            goalsDAO = getDao(GoalDAO.class);
+        return goalsDAO;
     }
 
     public static void outPutDatabase(Context context)
